Codec.NativeFromBinary: Clarify doc comment
The previous doc comment read:

On success, it returns the decoded datum, along with a new byte slice with the decoded bytes consumed, and a nil error value

I interpreted this as it returns the decoded datum and a byte slice containing the bytes that make up that datum (the decoded bytes). However, it actually returns the bytes that were not decoded. Attempt to reword it to make this clearer.