OCFReader has Block method that returns OCFBlock
Some use cases include using goavro to decompress an OCF file into
blocks, and allow the application to handle each block, possibly
skipping the block data or optionally decoding the data on demand.

This change add the Block method to OCFReader that returns an OCFBlock
instance that holds the decompressed Avro binary blob for the block,
and a count of the number of items in the block.

This is an experimental branch, and might be merged back into master
once feedback for this proposed API has been accepted.