logicalTypes: fix alternative encoding for decimal types
The alternative is an explicit codec decimal type, additionally, this
only worked for a small subset of values that were both valid native
and valid binary encoding. Since we want to stay in the native value
space, we just need to convert from native primative to native logical
type. In order to do this we just do an encode/decode step for decimal
logical types.